# docs/swagger.yaml
basePath: /v1
definitions:
  auth.AssignRoleRequest:
    properties:
      role_id:
        type: string
      user_id:
        type: string
    required:
    - role_id
    - user_id
    type: object
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
  auth.RoleDTO:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  auth.UserDTO:
    properties:
      id:
        type: string
      username:
        type: string
    type: object
  auth.UserInfoResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      profile:
        type: string
      roles:
        items:
          $ref: '#/definitions/auth.RoleDTO'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  pkgs.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: https://www.swagger.io/support
  description: This is a sample server for go-pg-demo server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Go-PG Demo API
  version: "1.0"
paths:
  /auth/assign-role:
    post:
      consumes:
      - application/json
      description: 为用户分配角色
      parameters:
      - description: 分配角色请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功，返回用户角色关联ID
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 分配角色
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 用户登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回访问令牌等信息
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 用户登录
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户信息
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.UserInfoResponse'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 获取当前用户信息
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 刷新访问令牌
      parameters:
      - description: 刷新令牌请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的访问令牌等信息
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.RefreshTokenResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "401":
          description: 无效的刷新令牌
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 刷新token
      tags:
      - auth
...
securityDefinitions:
  JWT:
    description: JWT token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"